refactor: move JobGuard to modules/jobs/guards
